move notifications manager test outside of qt subtree
