+++ tox.ini
@@ -85,7 +85,6 @@ commands_pre =
     # strictly only need to uninstall pytest-qt (which will raise without a backend)
     # the rest is for good measure
     headless: pip uninstall -y pytest-qt qtpy pyqt5 pyside2
-    !headless: pip install pydantic<1.8
 commands =
     !headless: pytest {env:PYTEST_PATH:} -v --color=yes --basetemp={envtmpdir} --cov-report=xml --cov={env:PYTEST_PATH:napari} {posargs}
     headless: pytest -v --color=yes --basetemp={envtmpdir} --ignore napari/_vispy --ignore napari/_qt --ignore napari/_tests {posargs}
